Modify init scripts.

+++ src/main/resources/kill.sh
@@ -1,7 +1,18 @@
 #!/bin/bash
-pid=`ps aux | grep ControlThread | awk '{print $2}'`
-echo "killing pid:" ${pid}
-kill -9 ${pid}
-pid=`ps aux | grep StatusThread | awk '{print $2}'`
-echo "killing pid:" ${pid}
-kill -9 ${pid}
+pid=`ps xauww | grep java | grep ControlThread | cut -d" " -f6`
+if [ -n "$pid" ]; then
+    echo "killing pid:" ${pid}
+    kill -9 ${pid}
+fi
+
+pid=`ps xauww | grep java | grep StatusThread | cut -d" " -f6`
+if [ -n "$pid" ]; then
+    echo "killing pid:" ${pid}
+    kill -9 ${pid}
+fi
+
+pid=`ps xauww | grep java | grep HvacMonitor | cut -d" " -f6`
+if [ -n "$pid" ]; then
+    echo "killing pid:" ${pid}
+    kill -9 ${pid}
+fi
